Created add book to shoppingCart logic backeend.
in CartItem service created method for adding book,user,and integer qty
in CartItemServiceImpl override method and i said list of cart items to
findByShoppingCart and for loop method for if user book is equals to
book cartItem book we are setting that same book in our database and qty
and subTotal BigDecimal(it's for precision).

If is the new book we are trying to add.
Then we instantiate the CartItem and set Book,CartItem,bookToCartItem,
copy pasted from top

return cartItem in the end.